requirements.txt
Django==3.2
django-classy-tags==2.0.0
django-cms==3.9.0
django-cors-headers==3.11.0
django-filer==2.1.2
django-formtools==2.3
django-js-asset==1.2.2
django-light==0.1.0.post3
django-mptt==0.13.4
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-snippet==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy_thumbnails==2.8
selenium==4.1.3
Pillow==9.0.0
psycopg2==2.9.3
mock==4.0.3
pipreqs
